refactor: classify components
