+++ src/components/ui/Input.tsx
@@ -1,9 +1,9 @@
 import React, { Fragment, useCallback, useMemo, useState } from "react";
 import { Combobox, Listbox, Transition } from "@headlessui/react";
 import { HiCheck, HiChevronUpDown } from "react-icons/hi2";
-import { availableTimezones } from "../utils/timezone";
+import { availableTimezones } from "../../utils/timezone";
 import { addMinutes, format, isAfter, isBefore, parse } from "date-fns";
-import { formatTime } from "../utils/utils";
+import { formatTime } from "../../utils/utils";
 
 interface SelectorProps {
   className?: string;
@@ -20,7 +20,7 @@ export const Selector: React.FC<SelectorProps> = ({
 }) => {
   return (
     <div
-      className={`rounded-input justify-stretch flex flex-row gap-1 rounded-lg p-0.5 text-sm ${
+      className={`rounded-input flex flex-row justify-stretch gap-1 rounded-lg p-0.5 text-sm ${
         className ?? ""
       }`}
     >
